Fixed compiler warning related to %08X and (size_t)
git-svn-id: svn+ssh://savannah.psi.ch/afs/psi.ch/project/meg/svn/mxml/trunk@31 e3eee304-8902-0410-aafd-c07b25ac10c4

+++ mxml.c
@@ -1984,10 +1984,10 @@ void mxml_debug_tree(PMXML_NODE tree, int level)
 
    for (i=0 ; i<level ; i++)
       printf("  ");
-   printf("Addr: %08X\n", (size_t)tree);
+   printf("Addr: %08X\n", (unsigned int)tree);
    for (i=0 ; i<level ; i++)
       printf("  ");
-   printf("Prnt: %08X\n", (size_t)tree->parent);
+   printf("Prnt: %08X\n", (unsigned int)tree->parent);
    for (i=0 ; i<level ; i++)
       printf("  ");
    printf("NCld: %d\n", tree->n_children);
